requirements/common.txt
aniso8601==3.0.0
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.4
PyJWT==1.6.1
python-dateutil==2.7.2
python-dotenv==0.8.2
pytz==2018.4
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
